Don't generate app_uri.txt if the `ruby_lsp_rails.server` is false
Because the addon client uses the file's existence to determine whether
to send the request, if the file is generated when the server
feature is disabled, the client would just spam the server with requests.

We also don't need to append the Rack app if Rails server is not invoked.